Removing the maximum limit on the requests dep
Having a maximum version that does not even exist today and thus cannot even
possibly break facepy is quite bad.

If requests decides to jump to 1.0 (or to any version greater than the
artificial maximum imposed in the setup), it does not necessarily mean that it
will break facepy.

In fact, in a project depending on a modern requests and facepy, installing
facepy after having installed a modern version of requests will probably remove
the new version of requests and will install the artificial maximum version
imposed by facepy.

If a modern version of requests breaks facepy, a bug will be filled and a fix
issued. Not the other way around...