Refactoring parse() into the SignedRequest constructor
The parse() method was not only parsing but building and returning a full
SignedRequest object. IMHO, this behaviour should go to __init__.

By refactoring, parse() is in charge of just parsing the signed_request and
returning its contents in a dict, while __init__ will know how to properly
build the full SignedRequest object.